Skip any test needing DB access if it's not accessible (for GitLab only; tests will still run on GitHub)
