+++ tests/integration/test_ephemeris_routes.py
@@ -1,41 +1,10 @@
 # ruff: noqa: E501, S101, F841
-import os
-from urllib.parse import urlparse
-
-import psycopg2
 import pytest
-import redis
+from tests.conftest import cannot_connect_to_services
 from tests.factories.satellite_factory import SatelliteFactory
 from tests.factories.tle_factory import TLEFactory
 
 from api.adapters.repositories.tle_repository import SqlAlchemyTLERepository
-
-
-def cannot_connect_to_services():
-    """Check if required services (Redis and Postgres) are available."""
-    # Check Redis
-    try:
-        r = redis.Redis(host="localhost", port=6379, db=0)
-        r.ping()
-    except redis.ConnectionError:
-        return "Can't connect to Redis"
-
-    # Check Postgres
-    try:
-        db_url = os.environ["SQLALCHEMY_DATABASE_URI"]
-        parsed = urlparse(db_url)
-        conn = psycopg2.connect(
-            dbname=parsed.path[1:],
-            user=parsed.username,
-            password=parsed.password,
-            host=parsed.hostname,
-            port=parsed.port,
-        )
-        conn.close()
-    except (psycopg2.OperationalError, KeyError):
-        return "Can't connect to PostgreSQL"
-
-    return False
 
 
 @pytest.mark.skipif(
