+++ tests/integration/test_tools_routes.py
@@ -1,6 +1,8 @@
 # ruff: noqa: E501, S101, F841
 from datetime import datetime
 
+import pytest
+from tests.conftest import cannot_connect_to_services
 from tests.factories.satellite_factory import SatelliteFactory
 from tests.factories.tle_factory import TLEFactory
 
@@ -8,6 +10,10 @@ from api.adapters.repositories import satellite_repository, tle_repository
 from api.entrypoints.extensions import db
 
 
+@pytest.mark.skipif(
+    cannot_connect_to_services(),
+    reason=cannot_connect_to_services() or "Services not available",
+)
 def test_get_tle_data(client, session):
     satellite = SatelliteFactory(sat_name="ISS")
 
@@ -20,6 +26,10 @@ def test_get_tle_data(client, session):
     assert response.status_code == 200
 
 
+@pytest.mark.skipif(
+    cannot_connect_to_services(),
+    reason=cannot_connect_to_services() or "Services not available",
+)
 def test_get_tle_data_no_match(client):
     response = client.get("/tools/get-tle-data/?id=ISS&id_type=name")
 
@@ -27,6 +37,10 @@ def test_get_tle_data_no_match(client):
     assert response.json == []
 
 
+@pytest.mark.skipif(
+    cannot_connect_to_services(),
+    reason=cannot_connect_to_services() or "Services not available",
+)
 def test_get_names_from_norad_id(client):
     satellite = SatelliteFactory(sat_number="25544")
     sat_repo = satellite_repository.SqlAlchemySatelliteRepository(db.session)
@@ -38,12 +52,20 @@ def test_get_names_from_norad_id(client):
     assert response.json[0]["norad_id"] == "25544"
 
 
+@pytest.mark.skipif(
+    cannot_connect_to_services(),
+    reason=cannot_connect_to_services() or "Services not available",
+)
 def test_get_names_from_norad_id_no_match(client):
     response = client.get("/tools/names-from-norad-id/?id=25544")
     assert response.status_code == 200
     assert response.json == []
 
 
+@pytest.mark.skipif(
+    cannot_connect_to_services(),
+    reason=cannot_connect_to_services() or "Services not available",
+)
 def test_get_norad_ids_from_name(client):
     satellite = SatelliteFactory(sat_name="ISS")
     sat_repo = satellite_repository.SqlAlchemySatelliteRepository(db.session)
@@ -55,12 +77,20 @@ def test_get_norad_ids_from_name(client):
     assert response.json[0]["name"] == "ISS"
 
 
+@pytest.mark.skipif(
+    cannot_connect_to_services(),
+    reason=cannot_connect_to_services() or "Services not available",
+)
 def test_get_norad_ids_from_name_no_match(client):
     response = client.get("/tools/norad-ids-from-name/?name=ISS")
     assert response.status_code == 200
     assert response.json == []
 
 
+@pytest.mark.skipif(
+    cannot_connect_to_services(),
+    reason=cannot_connect_to_services() or "Services not available",
+)
 def test_get_satellite_data(client):
     satellite = SatelliteFactory(sat_name="ISS")
     sat_repo = satellite_repository.SqlAlchemySatelliteRepository(db.session)
@@ -72,12 +102,20 @@ def test_get_satellite_data(client):
     assert response.json[0]["satellite_name"] == "ISS"
 
 
+@pytest.mark.skipif(
+    cannot_connect_to_services(),
+    reason=cannot_connect_to_services() or "Services not available",
+)
 def test_get_satellite_data_no_match(client):
     response = client.get("/tools/get-satellite-data/?id=ISS&id_type=name")
     assert response.status_code == 200
     assert response.json == []
 
 
+@pytest.mark.skipif(
+    cannot_connect_to_services(),
+    reason=cannot_connect_to_services() or "Services not available",
+)
 def test_get_tles_at_epoch(client, session):
     satellite = SatelliteFactory(
         sat_name="ISS", decay_date=None, has_current_sat_number=True
@@ -105,6 +143,10 @@ def test_get_tles_at_epoch(client, session):
     assert len(response.json[0]["data"]) == 0
 
 
+@pytest.mark.skipif(
+    cannot_connect_to_services(),
+    reason=cannot_connect_to_services() or "Services not available",
+)
 def test_get_tles_at_epoch_pagination(client, session):
     satellite = SatelliteFactory(
         sat_name="ISS", decay_date=None, has_current_sat_number=True
@@ -123,6 +165,10 @@ def test_get_tles_at_epoch_pagination(client, session):
     assert tles[0]["satellite_name"] == "ISS"
 
 
+@pytest.mark.skipif(
+    cannot_connect_to_services(),
+    reason=cannot_connect_to_services() or "Services not available",
+)
 def test_get_tles_at_epoch_optional_epoch_date(client, session):
     satellite = SatelliteFactory(
         sat_name="ISS", decay_date=None, has_current_sat_number=True
@@ -143,6 +189,10 @@ def test_get_tles_at_epoch_optional_epoch_date(client, session):
     assert tles[0]["satellite_name"] == "ISS"
 
 
+@pytest.mark.skipif(
+    cannot_connect_to_services(),
+    reason=cannot_connect_to_services() or "Services not available",
+)
 def test_get_tles_at_epoch_zipped(client, session):
     satellite = SatelliteFactory(
         sat_name="ISS", decay_date=None, has_current_sat_number=True
